+++ src/StreamTarget.php
@@ -31,10 +31,16 @@ final class StreamTarget extends Target
     /**
      * @param resource|string $stream A string stream identifier or a stream resource.
      * @param string[] $levels The {@see LogLevel log message levels} that this target is interested in.
+     * @param string|callable|null $contextFormat A context format for the log context output. See {@see Target::__construct()}.
+     * @param callable|null $stringConverter A PHP callable that converts a context value to a string. See {@see Target::__construct()}.
      */
-    public function __construct(private $stream = 'php://stdout', array $levels = [])
-    {
-        parent::__construct($levels);
+    public function __construct(
+        private $stream = 'php://stdout',
+        array $levels = [],
+        string|callable|null $contextFormat = null,
+        ?callable $stringConverter = null,
+    ) {
+        parent::__construct($levels, $contextFormat, $stringConverter);
     }
 
     protected function export(): void